refactor solidity deployment tools
